// src/pages/CustomerManagement/customerDetail/dxl/data.d.ts
export interface CustomerDetail {
  customerData: Partial<CustomerData>;
  requirementData: RequirementUser;
  contactUserData: ContactUserData[];
  reqTeamData: ProjectTeam[];
  // contractData:    ContractWrapper;
  reqLiteData: Partial<ReqLiteData>;
  followData: FollowData;
  cooperationData: CooperationData[];
  reserveData: ArriveData[];
}

export interface ArriveData {
  id: string;
  order_id: string;
  req_id: string;
  create_time: string;
  update_time: string;
  reserve_time: string;
  arrival_time: string;
  user_id: string;
  confirm_user_id: string;
  company_id: string;
  user_name: string;
  title: string;
  confirm_user_name: string;
}

export interface CooperationData {
  title: string	//标题
  comment: string	//内容
  create_user: string	//创建人
  create_time: string	//创建时间
}

export interface RequirementUser {
  my: RequirementData[];
  other: RequirementData[];
}

//有效单详情
export interface ReqLiteData {
  role: 0 | 1 | 2 | 3;  // 角色 0:属人,1:归属人,2:领导,3:协作人
  team_user: string;
  kefu: string;
  sale: string;
  next_contact_time: string;
  complaint_status: 0 | 1 | 2 | 3 | 4;  // 客诉审核状态  0表示正常无客诉单，1表示已有客诉单待审核处理，2表示跟进中，3表示投诉升级，4表示客诉单已处理完结可再次发起新的客诉
  complaint_id: string; // 客诉单id
  status: string;    //0无效单 1有效单 2客户合并
  phase: number;  // 有效单阶段：0已关单 1待跟进 2跟进中 3订金已付 4签订合同 5客户合并 6待回访 30无效
  owner_id: string;
  leads_id: string;
  req_id: string;
  associates: string;
  req_num: string;
  level: string;  //需求单评级
  level_txt: string; //评级描述
  category: string;
  top_category: string;

  qt_result: number;
  qt_result_txt: string;
  qt_result_reason: string;
}

export interface FollowData {
  showFollowButton: number;
  showFollowProductAndCoupon: number;
  showDistributedCompany: 0 | 1;  // 是否显示派发公司
  distributedCompany: any[];  // 派发公司列表
  followTab: FollowTab[];
}

export interface FollowTab {
  key: string;
  val: string;
}
// ...
export interface CustomerData {
  customerId: string;
  customerLevel: string;
  customerLevelText: string;
  channel: string;
  customerName: string;
  gender: string;
  genderText: string;
  identity: string;
  identityText: string;
  weddingDate: string;
  phone: string;
  weChat: string;
  liveCityInfo: CityInfo;
  liveAddress: string;
  workCityInfo: CityInfo;
  workAddress: string;
  contactTime: string;
  referrerName: string;
  likeCityInfo: CityInfo;
  budget: string;
  category: string;
  weddingStyle: string;
  weddingStyleText: string;
  deskNum: string;
  hotel: string;
  encryptPhone: string;
  nextContactTime: string;
  allotTime: string;
  createTime: string;
  createUser: string;

  referrerPhone: string;
  customerFollowStatus: string;
  editable: 0 | 1; // 0 不可编辑  1 可编辑

  status_txt: string;

  repeatStatus: 1 | 2 | 3;  //  客户重单状态 1:正常,2:客户合并,3:亲子单
  repeatAuditStatus: 0 | 1 | 2 | 3;  //  客户重单审核状态 0:正常无重单提审 1:待审核,2:驳回,3:通过
  status: string;   // 客户状态
  complaint_status: 0 | 1 | 2 | 3;  // 客诉审核状态
  
  weddingDateFrom:string;
  weddingDateEnd:string;

  allot_time:string; //客户划入时间
  record_user_id: string;  //(北京)提供人ID
  record_user_name: string; //(北京)提供人
}
// ...
export interface RequirementBean {
  id: string;
  req_num: string;
  category: string;
  budget: string;
  merchant: string;
  photo_style: string;
  photo_time: string;
  wedding_style: string;
  hotel: string;
  hotel_star: string;
  hotel_tables: string;
  hotel_hall: string;
  car_brand: string;
  car_series: string;
  car_num: string;
  car_time: string;
  est_arrival_time: string;
  status: string;
  status_txt: string;
  create_time: string;
  update_time: string;
  user_id: string;
  customer_id: string;
  user_name: string;
  remark: string;
  city_info: CityInfo;
  city_code: string;
  category_txt: string;
  phase: string;
}

export interface CityInfo {
  province; string;
  city: string;
  district: string;
  full: string;
  city_code: string;
}

export interface RequirementData {
  category_id: string;
  category_name: string;
  data: RequirementBean[];
}

/**
contactId	int	联系人id
userName	string	联系人姓名
phone	string	客户电话
identity	int	客户身份
identityText	string	客户身份
weChat	string	微信
occupation	string	职业
contactTime	int	方便联系时间
contactTimeText	string	方便联系时间
comment	string	备注
 */
export interface ContactUserData {
  customerId: string;
  contactId: string;
  userName: string;
  identity: string;
  identityText: string;
  phone: string;
  weChat: string;
  occupation: string;
  contactTime: string;
  contactTimeText: string;
  comment: string;
  encryptPhone: string;
}
// ...
export interface ProjectTeam {
  req_id: string;
  team_name: string;
  category: string;
  category_text: string;
  owner: string;
  owner_id: string;
  team_member: TeamMember[];
}

export interface TeamMember {
  id: string;
  team_user_id: string;
  team_key_name: string;
  team_user_name: string;
}